vigivagas_publico/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
Werkzeug==3.1.3
openpyxl==3.1.5
psycopg[binary]==3.2.3
gunicorn==23.0.0
requests==2.32.3
